==> 01_big-data.ipynb <==
# Big Data -- Big file not added

**-->** This notebook demonstrates what I did in terms of making the initial large data file smaller

- The data file was initially 14.2+ MB, with 60k rows. Through data cleaning, I was able to minimize the file to 1.4 MB.
- As a part of the cleaning, I changed data types to `category` to save memory.
- Additionally, I dropped columns I felt didn’t represent a user's lifestyle
    - `Essay` columns took up 3.4 MB
    - `Last_online`, `ethnicity`, `height`, `income`, `job`, `speaks`
- Lastly, I pickled the smaller subset of data

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

pd.set_option('display.max_columns', None)

# read in data
# initially reading in BIG data

cupid = pd.read_csv('data/okcupid_profiles.csv')

cupid.head(5)

# holy moly big data

cupid.info()

# only 60k rows, not the worst..

cupid.shape

----

## Initial EDA w/ Large Data File

cupid.isna().sum()

cupid['offspring'].value_counts()

cupid['education'].value_counts()

cupid['sex'].value_counts()

cupid['age'].value_counts()

cupid['body_type'].value_counts()

cupid['sign'].value_counts()

# drop rows where location != california

sum(cupid[cupid['location'].str.contains('california') == False]['location'].value_counts())

# group by what pet they like (likes/has dogs; likes/has cats) OR dislikes/no pets

cupid['pets'].value_counts()

# group by "laughing about it" or "not too serious about it" AS "doesn't matter"
# group by "somewhat serious about it" or "serious about it" AS "matters"
# all other values -- group by "matters"

cupid['religion'].value_counts()

# target variable?

cupid['orientation'].value_counts()

cupid['speaks'].value_counts()

----

## Make Smaller File

# pickle whole file, including essay columns? -- wayy too large still
# large file not saved

cupid.to_pickle('data/cupid.pkl')

cupid_check = pd.read_pickle('data/cupid.pkl')

cupid_check.info()

# drop essay columns
cupid_subset = cupid.iloc[:,:21]

cupid_subset.info()

# change object type to category
# referenced https://git.generalassemb.ly/DSI-322/7.01-big-data/blob/master/big-data-memory-post-class.ipynb

cat_cols = cupid_subset.drop(columns = 'height').columns
cupid_subset[cat_cols] = cupid_subset[cat_cols].astype('category')

cupid_subset.info()

cupid_subset['status'].value_counts()

# drop columns that weren't indicative to a user's lifestyle

cupid_subset.drop(columns = ['height', 'ethnicity', 'job', 'last_online', 'speaks', 'income', 'education'],inplace = True)

cupid_subset.info()

cupid_subset.to_pickle('data/subset_cupid.pkl')

cupid_subset_check = pd.read_pickle('data/subset_cupid.pkl')

cupid_subset_check.info()
