==> 02_initial-data-cleaning.ipynb <==
# Initial Data Cleaning

**-->** This notebook lays out my intial data cleaning process where I:
- Further dropped columns `education` and `sign`, as I felt that they weren't indicative of recommending a profile to a user
- Filled missing values
- Filtered the data on certain features
    - status: 'single' or 'available'
    - location: California only
    - age: filtered out ages 109 + 110, as they were significantly older than the dataset

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

# read in data from pickle file

cupid_df = pd.read_pickle('data/subset_cupid.pkl')

cupid_df.head()

#### Further filtering data
- Dropping 'education' and 'sign' features, as I personally felt that the values were odd ('education') or not indicative of recommending a partner ('sign')
- Filtering 'status' to only 'single' or 'available', since we're recommending lovers and 'married' or 'seeing someone' already have a person of interest, assuming relationships aren't polyamorous

# drop education + sign -- useless / not indicative, in my opinion

cupid_df.drop(columns = ['education', 'sign'], inplace = True)

cupid_df.shape

cupid_df.info()

cupid_df['status'].value_counts()

# filter for only those who are 'single' or 'available'

cupid = cupid_df[(cupid_df['status'] == 'single') | (cupid_df['status'] == 'available')]

# dropped ~2k rows

cupid.shape

#### Null Values

# check for nulls

cupid.isna().sum()

# impute missing values

cupid['body_type'].fillna('rather not say', inplace = True)
cupid['diet'].fillna('anything', inplace = True)
cupid['drinks'].fillna('not at all', inplace = True)
cupid['drugs'].fillna('never', inplace = True)
cupid['offspring'].fillna("doesn't have kids", inplace = True)
cupid['pets'].fillna('dislikes dogs and dislikes cats', inplace = True)
cupid['religion'].fillna('atheism', inplace = True)
cupid['smokes'].fillna('no', inplace = True)

# ensure null values are handled

cupid.isna().sum()

# filter for age
# entries where age is 109 + 110

cupid['age'].value_counts()

# change to data type int
cupid['age'] = cupid['age'].astype('int32')

# filter ages 109 + 110
cupid = cupid[cupid['age'] < 109]

# filter for location only in california
cupid = cupid[cupid['location'].str.contains('california')]

cupid['location'].value_counts()

# pickle the initial cleaned data
cupid.to_pickle('data/clean_cupid.pkl')

cupid_check = pd.read_pickle('data/clean_cupid.pkl')

cupid_check.info()
